Various changes

+++ src/alsa_input.h
@@ -17,5 +17,5 @@
  */
 
 extern int open_alsa_input(char *input_card, unsigned int sample_rate, size_t buf_size);
-extern int read_alsa_input(float *buffer);
+extern int read_alsa_input(short *buffer);
 extern int close_alsa_input();
